+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 - `ReactCommands` now requires reactive types implement `ReactComponent` and `ReactResource`.
 - `EventRevokeToken` was removed in favor of unified `RevokeToken`s for all reactors.
 - Named systems are now mapped to both the input id and the system type, instead of just the input id. This allows the internal named systems cache to not be parameterized by the system type, which makes it easier to access.
-- `InteractiveElementBuilder` callbacks no longer take the cursor world position as input. Use `CursorWorldPos` system parameter to access the current cursor instead.
+- `InteractiveElementBuilder` callbacks no longer take the cursor world position as input. Use `CursorPos` system parameter to access a cursor's position instead.
 
 
 ## [0.2.0]